changed stepangle from int to float to get more accurate target calculation
if using high microstepping the stepangel gets too small to perform
integer calculations, so i changed it to float. makes everything a
little slower (can get up to 18 kHz with 200 steps/mm an 200 mm/s) but
gives more precision.

+++ Standard_Mechaduino/Mechaduino_01/State.h
@@ -31,7 +31,7 @@ extern int ITerm_max;
 //---- Step settings -----
 extern const int counts_per_revolution; // encoder counts for 360 degrees
 extern const float angle_per_step;      // only needed for calibration routine by now
-extern const int stepangle;             // angle of one step as int
+extern const float stepangle;             // angle of one step as int
 extern const int PA;                    // angle of one fullstep
 extern volatile int step_target;        // target as step gets incremented if an step is received
